Cell pooling.
