+++ src/PrettyPrompt/Rendering/ScreenArea.cs
@@ -1,0 +1,28 @@
+﻿#region License Header
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+#endregion
+
+using System;
+using PrettyPrompt.Consoles;
+
+namespace PrettyPrompt;
+
+/// <summary>
+/// An area of the screen that's being rendered at a coordinate.
+/// This is conceptually a UI pane, rasterized into characters.
+/// </summary>
+internal sealed record ScreenArea(ConsoleCoordinate Start, Row[] Rows, bool TruncateToScreenHeight = true) : IDisposable
+{
+    public void Dispose()
+    {
+        foreach (var row in Rows)
+        {
+            row.Dispose();
+        }
+#if DEBUG
+        Array.Clear(Rows, 0, Rows.Length);
+#endif
+    }
+}
